make RAC select handle ensuring press scaling still works even when the overlay is opened on mousedown
